feat: Remove utils.js and add functions to their respective components
