+++ src/components/Pokemon/Pokemon.js
@@ -2,10 +2,23 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import './Pokemon.css';
-import { formatDexNum } from '../../utils/helpers';
 
 const Pokemon = ({ dexNum, name }) => {
   const spriteSrc = `https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/${dexNum}.png`
+
+  const formatDexNum = (num) => {
+    const strNum = num.toString();
+  
+    if (strNum.length === 1) {
+      return `#000${strNum}`
+    } else if (strNum.length === 2) {
+      return `#00${strNum}`
+    } else if (strNum.length === 3) {
+      return `#0${strNum}`
+    } else {
+      return `#${strNum}`
+    }
+  }
 
   return (
     <figure>
